Fix: Integration of new renderstate into DX render manager. Few warning fixes in VS2005

+++ Sources/Internal/FileSystem/DynamicMemoryFile.cpp
@@ -111,7 +111,7 @@ uint32 DynamicMemoryFile::Read(void * pointerToData, uint32 dataSize)
 	}
 	
 	int32 realReadSize = dataSize;
-	int32 size = data.size();
+	int32 size = (int32)data.size();
 	if (currentPtr + realReadSize > size)
 	{
 		realReadSize = size - currentPtr;
